+++ lib/src/kick_chat.dart
@@ -1,5 +1,8 @@
 import 'dart:async';
+import 'dart:io';
 
+import 'package:api_7tv/api_7tv.dart';
+import 'package:fk_user_agent/fk_user_agent.dart';
 import 'package:flutter/foundation.dart';
 import 'package:kick_chat/kick_chat.dart';
 import 'package:web_socket_channel/io.dart';
@@ -12,6 +15,7 @@ class KickChat {
   StreamSubscription? _streamSubscription;
 
   KickUser? userDetails;
+  List seventvEmotes = [];
 
   Function()? onDone;
   final Function? onError;
@@ -26,14 +30,27 @@ class KickChat {
     this.onError,
   });
 
+  static Future init() async {
+    if(!Platform.isAndroid && !Platform.isIOS) {
+      return;
+    }
+    await FkUserAgent.init();
+  }
+
   Future<void> connect() async {
     userDetails = await KickApi.getUserDetails(username);
     if (userDetails == null) {
       return;
     }
 
+    // get channel 7tv emotes
+    List result = await SeventvApi.getKickChannelEmotes(userDetails!.userId.toString()) ?? [];
+    seventvEmotes.addAll(result);
+
     _webSocketChannel = IOWebSocketChannel.connect(
         "wss://ws-us2.pusher.com/app/{$pushKey}?protocol=7&client=js&version=7.6.0&flash=false");
+    _webSocketChannel?.sink.add(
+        '{"event":"pusher:subscribe","data":{"auth":"","channel":"channel.${userDetails!.id}"}}');
     _webSocketChannel?.sink.add(
         '{"event":"pusher:subscribe","data":{"auth":"","channel":"chatrooms.${userDetails!.chatRoom.id}.v2"}}');