+++ lib/src/kick_event.dart
@@ -1,6 +1,15 @@
 enum TypeEvent {
   message,
   followersUpdated,
+  streamHostEvent,
+  subscriptionEvent,
+  chatroomUpdatedEvent,
+  userBannedEvent,
+  chatroomClearEvent,
+  giftedSubscriptionsEvent,
+  pinnedMessageCreatedEvent,
+  pollUpdateEvent,
+  messageDeletedEvent,
 }
 
 class KickEvent {