+++ CHANGELOG.md
@@ -1,3 +1,47 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+Initial release
+
+## 1.0.1
+
+Add exports
+
+## 1.0.2
+
+User agents for Kick API
+
+## 1.0.3
+
+Fix stream host model
+
+## 1.0.4
+
+add an init for FkUserAgent
+
+## 1.0.5
+
+Use userAgent instead of WebviewUserAgent
+
+## 1.0.6
+
+Fix user banned
+
+## 1.0.7
+
+Fix user banned type
+
+# 1.0.8
+
+fix chatroom clear type
+
+# 1.1.0
+
+Get 7TV emotes
+
+## 1.1.1
+
+Bump depedencies
+
+## 1.2.0
+
+Bump depedencies